Move mdp decoding to its own module to make it easier to reuse with other capture formats
